Combine calendar dialogs and support timezone
