+++ openc3-cosmos-init/plugins/packages/openc3-cosmos-tool-scriptrunner/src/tools/ScriptRunner/ScriptRunner.vue
@@ -342,7 +342,7 @@
       v-if="inputMetadata.show"
       v-model="inputMetadata.show"
       :events="inputMetadata.events"
-      :utc="false"
+      :time-zone="timeZone"
       new-metadata
       @close="inputMetadata.callback"
     />
